feat(pix1D): rename files
