+++ mon-pix/app/utils/PIX1D/gestion_fichier/fichier-level-1.js
@@ -1,8 +1,8 @@
-export default class CampaignLevel1 {
+export default class FichierLevel1 {
   constructor(masteryPercentage) {
     this.masteryPercentage = masteryPercentage;
   }
   next() {
-    return 'NTNSFE441';
+    return 'GRFSLV922';
   }
 }
