+++ mon-pix/app/utils/PIX1D/gestion_fichier/fichier-level-2.js
@@ -1,12 +1,12 @@
-export default class CampaignLevel2 {
+export default class FichierLevel2 {
   constructor(masteryPercentage) {
     this.masteryPercentage = masteryPercentage;
   }
   next() {
     if (this.masteryPercentage < 70) {
-      return 'KLWLLY575';
+      return 'SDDPGG528';
     } else {
-      return 'SOURISPAD';
+      return 'DOCUMENTS';
     }
   }
 }
